+++ test_classes.py
@@ -1,4 +1,4 @@
-from unit_converter.api import Converter
+from src.unit_converter.api import Converter
 
 
 converter = Converter()